UI BUG Fix 5.	Add new profile a.	Company auto generation password shouldn’t be available in the first loading. 6.	Job List a.	“Company Name” should be in the job list table. b.	Table data should be left aligned. c.	Count shouldn’t be in the “Job List” screen.

+++ NIBM_Job_Portal/Controllers/HomeController.cs
@@ -117,7 +117,7 @@ namespace NIBM_Job_Portal.Controllers
                 company.Website = model.Website;
                 company.IsEnable =(int) CompanyStatus.Enable;
                 company.Contact_No = model.Contact_No;
-                company.IndustryId = model.IndustryId;
+                company.IndustryId = (int)model.IndustryId;
